// src/enums/locatorBy.ts
// Copyright 2021 TestProject (https://testproject.io)
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//   http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

/**
 * Supported element locator strategies.
 *
 * @enum {string}
 */
const enum LocatorBy {
  ID = 'id',
  XPATH = 'xpath',
  LINK_TEXT = 'link text',
  PARTIAL_LINK_TEXT = 'partial link text',
  NAME = 'name',
  TAG_NAME = 'tag name',
  CLASS_NAME = 'class name',
  CSS_SELECTOR = 'css selector',
}

export default LocatorBy;